Split Selector into its own file
